RR-732 - Rename file to dynamicAriaTextResolver
