+++ server/routes/induction/update/qualificationLevelUpdateController.ts
@@ -2,7 +2,7 @@ import { NextFunction, Request, RequestHandler, Response } from 'express'
 import QualificationLevelController from '../common/qualificationLevelController'
 import validateQualificationLevelForm from './qualificationLevelFormValidator'
 import { addPage, getNextPage, getPreviousPage } from '../../pageFlowQueue'
-import getDynamicBackLinkAriaText from '../dynamicAriaText'
+import getDynamicBackLinkAriaText from '../dynamicAriaTextResolver'
 
 /**
  * Controller for the Update of the Qualification Level screen of the Induction.
